+++ mobile/android/android-components/components/browser/engine-gecko/src/test/java/mozilla/components/browser/engine/gecko/GeckoEngineSessionTest.kt
@@ -315,10 +315,6 @@ class GeckoEngineSessionTest {
                 override fun onCookieBannerChange(status: CookieBannerHandlingStatus) {
                     cookieBanner = status
                 }
-                override fun onProductUrlChange(isProductUrl: Boolean) {
-                    displaysProduct = isProductUrl
-                }
-
                 override fun onTranslatePageChange() {
                     translationsProcessing = false
                 }