+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/action/BrowserAction.kt
@@ -854,14 +854,6 @@ sealed class ContentAction : BrowserAction() {
     data class CheckForFormDataExceptionAction(val tabId: String, val throwable: Throwable) : ContentAction()
 
     /**
-     * Updates the [ContentState.isProductUrl] state for the non private tab with the given [tabId].
-     */
-    data class UpdateProductUrlStateAction(
-        val tabId: String,
-        val isProductUrl: Boolean,
-    ) : ContentAction()
-
-    /**
      * Inform that the tab with [tabId] started rendering a pdf.
      */
     data class EnteredPdfViewer(