+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/reducer/ContentStateReducer.kt
@@ -314,15 +314,6 @@ internal object ContentStateReducer {
             -> {
                 throw IllegalStateException("You need to add SessionPrioritizationMiddleware. ($action)")
             }
-            is ContentAction.UpdateProductUrlStateAction -> {
-                updateContentState(state, action.tabId) {
-                    if (it.private) {
-                        it
-                    } else {
-                        it.copy(isProductUrl = action.isProductUrl)
-                    }
-                }
-            }
             is ContentAction.EnteredPdfViewer -> {
                 updateContentState(state, action.tabId) {
                     it.copy(isPdf = true)